refactor(orderflow): split projection.rs and config.rs into owned siblings
Move the read model, the print tiers and the budget fold of the heatmap
projection out of `projection.rs` into `projection/{model,tiers,fold}.rs`,
and the bubble and live-lane config families out of `config.rs` into
`config/{bubbles,lane}.rs`. Every body moves unchanged; the only widenings
are the `pub(super)` marks a root now needs to name a sibling's item, and
each root re-exports every moved `pub` item so no public path changes and
no consumer is edited. Every file is now under the 1,500-production-line
threshold: `projection.rs` 1,850 -> 836, `config.rs` 1,523 -> 514, the
largest sibling 556.

The two signed entries leave `size-baseline.txt` and `!budget` falls by
exactly the two ceilings, 20,025 -> 16,652. Test files gain `use` lines
only; every golden, parity and determinism test in `orderflow` and
`engine` passes unchanged.

Campaign child of #367 (S8); closes on integration: #375.
