+++ crates/orderflow/src/config.rs
@@ -1,10 +1,25 @@
 //! Sanitized runtime configuration for the order-book heatmap.
 
 use rust_decimal::Decimal;
-use rust_decimal::prelude::FromPrimitive as _;
-use serde::{Deserialize, Serialize};
 
-use super::history::TapeAge;
+mod bubbles;
+mod lane;
+
+pub use bubbles::{
+    BubbleRenderMode, BubbleSizeReference, BubbleStyle, ConsumptionMark, DEFAULT_BUBBLE_MAX_RADIUS,
+    DEFAULT_BUBBLE_MIN_RADIUS, DEFAULT_BUBBLE_OPACITY, DEFAULT_DETAIL_MIN_RADIUS,
+    DEFAULT_FRONT_LENGTH_SCALE, DEFAULT_LABEL_MIN_RADIUS, DEFAULT_LABEL_MIN_RADIUS_SHARE,
+    DEFAULT_READABLE_MIN_RADIUS, DEFAULT_SPHERE_HIGHLIGHT, DEFAULT_SPHERE_SHADING, GOLDEN_ANGLE,
+    INV_PHI, INV_PHI_2, INV_PHI_3, MAX_BUBBLE_MAX_RADIUS, MAX_BUBBLE_MIN_RADIUS,
+    MAX_READABLE_MIN_RADIUS, MIN_BUBBLE_MAX_RADIUS, PHI,
+};
+pub use lane::{
+    DEFAULT_LIVE_LANE_RADIUS_SCALE, DEFAULT_LIVE_LANE_SHARE, DEFAULT_LIVE_LANE_ZOOM,
+    LANE_WINDOW_PRESETS_MS, LaneWindow, LiveLaneStyle, MAX_LIVE_LANE_RADIUS_SCALE,
+    MAX_LIVE_LANE_SHARE, MAX_LIVE_LANE_WINDOW_MS, MAX_LIVE_LANE_ZOOM, MIN_LIVE_LANE_RADIUS_SCALE,
+    MIN_LIVE_LANE_SHARE, MIN_LIVE_LANE_WIDTH_PX, MIN_LIVE_LANE_WINDOW_MS, MIN_LIVE_LANE_ZOOM,
+    format_window_ms, lane_lag_label, lane_window_label, same_lane_window,
+};
 
 /// Shortest history window accepted by the UI.
 pub const MIN_RETENTION_MS: i64 = 1_000;
@@ -54,148 +69,6 @@ pub const MAX_BUBBLE_REGION_MS: i64 = 10_000;
 pub const DEFAULT_LIQUIDITY_CORRELATION_MS: i64 = 250;
 /// Safe upper bound for depth/aggression correlation.
 pub const MAX_LIQUIDITY_CORRELATION_MS: i64 = 10_000;
-/// The golden ratio, φ.
-///
-/// The aggression layer's proportions are steps on this one ladder rather than
-/// a scatter of independently tuned numbers. Radius is *area*-proportional to
-/// quantity, so a φ step in radius is a φ² ≈ 2.6× step in quantity — "roughly
-/// two and a half times the print" is a size class a flow trader already reads
-/// in, which is what makes the ladder a measuring stick and not an ornament.
-///
-/// It is used where a *proportion* is being expressed. It is deliberately not
-/// used for pixel floors (below ~3px the display grid rules, and a 0.38px
-/// stroke is a rounding error) nor for alphas, which are perceptual and tuned
-/// against the heat ramp's actual luminance.
-pub const PHI: f32 = 1.618_034;
-/// 1/φ ≈ 0.618 — see [`PHI`]. Each rung is derived from the last rather than
-/// written out, so the ladder cannot drift apart one literal at a time.
-pub const INV_PHI: f32 = 1.0 / PHI;
-/// 1/φ² ≈ 0.382 — see [`PHI`].
-pub const INV_PHI_2: f32 = INV_PHI / PHI;
-/// 1/φ³ ≈ 0.236 — see [`PHI`].
-///
-/// The bottom rung of the ladder, `1/φ⁴`, is deliberately not a constant of
-/// its own: nothing computes with it at runtime — the smallest radius is a
-/// stored default, not a derived value — so it would be a name with no
-/// caller. It is written as `INV_PHI_3 / PHI` where it is needed.
-pub const INV_PHI_3: f32 = INV_PHI_2 / PHI;
-/// The golden angle, 360°/φ² ≈ 137.5°, in radians.
-///
-/// The sweep a fully consuming print's crown reaches. Being 1/φ² of the rim
-/// rather than all of it is what makes the crown structurally incapable of
-/// closing into a second circle around the bubble — the exact ambiguity the
-/// old closed impact ring created with the disc's own edge.
-pub const GOLDEN_ANGLE: f32 = std::f32::consts::TAU * INV_PHI_2;
-/// Default alpha applied to aggression bubbles.
-///
-/// High enough that a bubble occludes rather than tints: translucency over a
-/// coloured heat cell drags the fill toward the heat's own hue, and occlusion
-/// is the strongest cue for counting overlapping prints. Short of opaque, so a
-/// bubble still admits it is sitting over the book.
-pub const DEFAULT_BUBBLE_OPACITY: f32 = 0.9;
-/// Default on-screen radius, in points, of the largest aggression bubble.
-pub const DEFAULT_BUBBLE_MAX_RADIUS: f32 = 15.0;
-/// Smallest accepted maximum bubble radius.
-pub const MIN_BUBBLE_MAX_RADIUS: f32 = 4.0;
-/// Largest accepted maximum bubble radius.
-pub const MAX_BUBBLE_MAX_RADIUS: f32 = 48.0;
-/// Largest accepted radius for the smallest drawn bubble.
-pub const MAX_BUBBLE_MIN_RADIUS: f32 = 12.0;
-/// Default radius of the smallest drawn print: the bottom rung of the φ
-/// ladder, `max / φ⁴`. See [`PHI`].
-///
-/// Written as a literal, not as `DEFAULT_BUBBLE_MAX_RADIUS * INV_PHI_4`, for
-/// one mechanical reason: the presets file stores floats to
-/// [`SERIALIZED_FLOAT_PLACES`] decimals, so a default carrying more precision
-/// than that would come back different from a save-and-reload and every
-/// preset round trip would drift. The rungs are therefore the exact ladder
-/// rounded to what the file can hold — asserted by
-/// `the_default_radii_are_rungs_of_one_golden_ladder`.
-pub const DEFAULT_BUBBLE_MIN_RADIUS: f32 = 2.1885;
-/// Default radius at which a bubble can afford its dressing — shading and the
-/// separator hair. The `max / φ³` rung.
-///
-/// It lands where hue starts working: below roughly 3.5px the eye's chromatic
-/// channel is too low-pass to separate a green speck from a red one, so that
-/// is exactly the size at which shading a bubble by side begins to pay. Under
-/// it a print is one flat disc, which is also what keeps the per-frame
-/// tessellation budget flat on a fast tape.
-pub const DEFAULT_DETAIL_MIN_RADIUS: f32 = 3.541;
-/// Default radius, in pixels, below which a bubble is treated as unreadable
-/// on its own: small enough to be a routine print, large enough that side
-/// colour and the side nudge actually register at a glance. The `max / φ²`
-/// rung, and the floor the consumption crown and the pie split gate on.
-pub const DEFAULT_READABLE_MIN_RADIUS: f32 = 5.7295;
-/// Default half-length of the vertical consumption front, as a multiple of the
-/// radius: `1/φ`, at the precision the presets file stores. Only
-/// [`ConsumptionMark::Front`] reads it.
-pub const DEFAULT_FRONT_LENGTH_SCALE: f32 = 0.618;
-/// Smallest radius that gets a label, as a fraction of
-/// [`DEFAULT_BUBBLE_MAX_RADIUS`].
-///
-/// Not a φ rung, and deliberately not dressed up as one: the rule is "the top
-/// of the radius range", because laying out text is the most expensive thing
-/// a bubble can ask for and only the largest prints have room to hold it.
-pub const DEFAULT_LABEL_MIN_RADIUS_SHARE: f32 = 0.9;
-/// Smallest radius that gets a label, in pixels — see
-/// [`DEFAULT_LABEL_MIN_RADIUS_SHARE`].
-pub const DEFAULT_LABEL_MIN_RADIUS: f32 =
-    DEFAULT_BUBBLE_MAX_RADIUS * DEFAULT_LABEL_MIN_RADIUS_SHARE;
-/// Largest accepted readability floor.
-pub const MAX_READABLE_MIN_RADIUS: f32 = 32.0;
-/// Default edge darkening of a sphere-rendered bubble.
-///
-/// Enough contrast that two piled bubbles keep a boundary, and no more:
-/// counting overlapping discs needs only a soft edge, while a hard one reads
-/// as a black outline drawn around every print on the tape.
-pub const DEFAULT_SPHERE_SHADING: f32 = 0.34;
-/// Default highlight strength of a sphere-rendered bubble.
-///
-/// A specular cue, not a second colour. Pushed harder, the sell side's core
-/// washes out to pink — the loudest thing on the canvas and the one that makes
-/// the layer read as decoration rather than as instrumentation.
-pub const DEFAULT_SPHERE_HIGHLIGHT: f32 = 0.18;
-/// Default share of the chart width taken by the live lane. Enough room for
-/// the tape to be read as a tape, with two thirds of the chart left for the
-/// history it is read against.
-pub const DEFAULT_LIVE_LANE_SHARE: f32 = 0.35;
-/// Narrowest live lane accepted, as a share of the chart.
-pub const MIN_LIVE_LANE_SHARE: f32 = 0.05;
-/// Widest live lane accepted. The tape may take half the chart and no more:
-/// past that there is no history left to read the tape against.
-pub const MAX_LIVE_LANE_SHARE: f32 = 0.5;
-/// Narrowest lane the layout will draw, in pixels. A floor rather than a
-/// setting: below this the band stops being a tape whatever share was asked
-/// for. Capped by the chart itself, so a tiny window is still split by share.
-pub const MIN_LIVE_LANE_WIDTH_PX: f32 = 24.0;
-/// Default zoom of the lane's time window: one typical bar's worth of market
-/// time, the same window the lane showed before the knob existed.
-pub const DEFAULT_LIVE_LANE_ZOOM: f32 = 1.0;
-/// Most zoomed-out lane accepted — four times the typical bar duration in the
-/// same band, so prints crowd together and cluster into fewer, bigger marks.
-pub const MIN_LIVE_LANE_ZOOM: f32 = 0.25;
-/// Most zoomed-in lane accepted — an eighth of it, so prints run across the
-/// band fast and far apart.
-pub const MAX_LIVE_LANE_ZOOM: f32 = 8.0;
-/// Least market time a zoomed-in lane will show. Below an instant there is no
-/// tape left, whatever the zoom asked for.
-pub const MIN_LIVE_LANE_WINDOW_MS: i64 = 200;
-/// Most market time a zoomed-out lane will show. A tape is the recent past;
-/// past a quarter of an hour the chart's own history says it better.
-pub const MAX_LIVE_LANE_WINDOW_MS: i64 = 900_000;
-/// The fixed tape windows the lane's menu offers, in exchange milliseconds.
-///
-/// Round durations a trader already thinks in, not a sample of the accepted
-/// range: the point of a preset is that it is chosen without reading a number.
-/// Anything else is reachable through the custom entry, which accepts the
-/// whole [`MIN_LIVE_LANE_WINDOW_MS`]..=[`MAX_LIVE_LANE_WINDOW_MS`] band.
-pub const LANE_WINDOW_PRESETS_MS: [i64; 5] = [15_000, 30_000, 60_000, 120_000, 300_000];
-/// Default multiplier applied to the bubble radii inside the live lane.
-pub const DEFAULT_LIVE_LANE_RADIUS_SCALE: f32 = 1.0;
-/// Bounds accepted for the live lane's radius multiplier.
-pub const MIN_LIVE_LANE_RADIUS_SCALE: f32 = 0.25;
-/// See [`MIN_LIVE_LANE_RADIUS_SCALE`].
-pub const MAX_LIVE_LANE_RADIUS_SCALE: f32 = 4.0;
 
 /// Renderer-only price grouping layered over the exact capture buckets.
 ///
@@ -270,893 +143,11 @@ impl IntensityMode {
     }
 }
 
-/// How the quantity that maps to a full-size bubble is chosen.
-///
-/// The automatic modes measure the whole recorded tape, never the visible
-/// window: zoom decides what is on screen, not what a quantity means, so the
-/// scale only moves when the recording itself does — a new print beyond the
-/// old reference, or retention trimming the floor. `VisibleP99` keeps one
-/// outlier sweep from shrinking every other print to a dot, at the cost of
-/// making the top 1% all render at the maximum radius. `VisibleMax` restores
-/// a strict ordering among recorded prints (clusters a zoomed-out view merges
-/// past it saturate at full size); `Fixed` pins the scale so bubble size
-/// means the same thing across sessions and symbols. The `Visible*` names are
-/// the wire format shipped presets already use; what they measure stopped
-/// being the viewport the day zoom stopped being allowed to rescale a print.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum BubbleSizeReference {
-    /// 99th percentile of the recorded clustered prints.
-    #[default]
-    VisibleP99,
-    /// The largest recorded clustered print.
-    VisibleMax,
-    /// An explicit quantity, [`BubbleStyle::size_reference_quantity`].
-    Fixed,
-}
-
-/// How a print that ate resting liquidity says so.
-///
-/// The signal has to survive three things at once: a dense tape where prints
-/// overlap, a canvas that is nearly black, and a heat map underneath that is
-/// the only layer carrying liquidity. That rules out any mark that works by
-/// *subtraction* — a dark stroke is invisible over the canvas and a hole
-/// punched in the book everywhere else — and it rules out any mark that
-/// crosses the disc, whose area is the quantity reading.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ConsumptionMark {
-    /// An open arc outside the rim, on the side of the book that was eaten:
-    /// buy aggression lifts the ask, so its crown sits above; sell aggression
-    /// hits the bid and wears it below.
-    ///
-    /// Arc *length* carries how much of the print matched resting liquidity —
-    /// the one channel the eye reads ordinally without a reference beside it —
-    /// from [`INV_PHI_2`] of the [`GOLDEN_ANGLE`] at a nibble to the whole
-    /// golden angle at a full sweep. Because it never closes and never crosses
-    /// the disc, it cannot be confused with the bubble's own rim, and stacked
-    /// crowns at one price row read as the nested brackets that absorption
-    /// actually is.
-    #[default]
-    Crown,
-    /// The vertical line through the bubble that shipped before the crown.
-    ///
-    /// Kept because a chart is someone's muscle memory and a preset that asked
-    /// for it should still get it — not because it reads well on a fast tape,
-    /// where it slices every disc it marks.
-    Front,
-    /// No consumption mark at all. The heat map to the right of the print
-    /// still shows whether the level refilled, which is the honest record.
-    None,
-}
-
-impl ConsumptionMark {
-    /// Whether this mark draws the vertical front.
-    #[must_use]
-    pub fn is_front(self) -> bool {
-        matches!(self, Self::Front)
-    }
-}
-
-/// How a bubble's fill is painted.
-///
-/// `Flat` is the classic solid disc. `Sphere` shades each bubble like a ball
-/// lit from the upper left — an offset highlight over a darkened rim — so
-/// overlapping prints on a dense tape keep a visible boundary instead of
-/// merging into one solid blob. Purely visual: geometry, clustering and
-/// liquidity association are identical in both modes.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum BubbleRenderMode {
-    /// Solid 2D disc.
-    #[default]
-    Flat,
-    /// Shaded 3D-looking sphere.
-    Sphere,
-}
-
-/// Decimal places kept when a bubble float is written to the presets file.
-///
-/// Four is past the precision any pixel radius or alpha carries, and well
-/// inside what `f32` represents exactly enough to survive a round trip.
-const SERIALIZED_FLOAT_PLACES: i32 = 4;
-
-/// Serialize an `f32` as a short decimal.
-///
-/// TOML floats are `f64`, so an `f32` promoted straight through prints its
-/// exact binary expansion: `0.78` is written back as `0.7799999713897705`.
-/// Presets are tracked in git precisely so a look can be reviewed and rolled
-/// back like code, and a diff of noise defeats that — this is the write path
-/// for every visual float, so rounding here fixes the file for all of them.
-fn serialize_short_f32<S>(value: &f32, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    let scale = 10_f64.powi(SERIALIZED_FLOAT_PLACES);
-    let rounded = (f64::from(*value) * scale).round() / scale;
-    serializer.serialize_f64(rounded)
-}
-
-/// Everything the "aggression bubbles" panel owns: bubble geometry, colour,
-/// labels and the two marks a bubble draws when it ate resting liquidity — the
-/// vertical consumption front (the "risco") and the trail that leaks into the
-/// consumed side (the "rastro").
-///
-/// Every field here is display-only. [`min_quantity`](Self::min_quantity) and
-/// the size reference reach the projection, but only to decide which bubbles
-/// are drawn and how large — never what is captured, retained, or associated
-/// with an L2 reduction.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct BubbleStyle {
-    /// Radius of the smallest drawn print, in pixels.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub min_radius: f32,
-    /// Radius of a full-size print, in pixels. Area stays proportional to
-    /// quantity between the two.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub max_radius: f32,
-    /// Alpha of the bubble fill.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub opacity: f32,
-    /// Rim stroke width; zero draws no rim.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub outline_width: f32,
-    /// Soft halo drawn behind the fill, as a fraction of full alpha.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub halo_strength: f32,
-    /// Bubbles below this radius are drawn as a plain dot (no halo, rim or
-    /// impact ring). Raising it trades detail for frame time on a fast tape.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub detail_min_radius: f32,
-    /// Radius, in pixels, below which a bubble cannot be read on its own.
-    ///
-    /// The readability floor both the dust merge and
-    /// [`hollow_small_buys`](Self::hollow_small_buys) gate on. Deliberately
-    /// *not* [`detail_min_radius`](Self::detail_min_radius): that one decides
-    /// how much dressing a bubble can afford, and a sphere-heavy look sets it
-    /// low on purpose — the "3d spheres" and "dense tape btc" presets put it
-    /// at or below `min_radius`, which would leave nothing to fold.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub readable_min_radius: f32,
-    /// Whether buy prints below [`readable_min_radius`](Self::readable_min_radius)
-    /// are drawn as an open ring instead of a solid dot.
-    ///
-    /// Off by default. The intent was that shape survives where colour does
-    /// not, but the ring spends most of the disc's area to say it: at 2px a
-    /// green ring around a 22%-alpha interior carries *less* green than the
-    /// dot it replaced, so the side became harder to read, not easier — and
-    /// the two sides ended up drawn in visibly different media. Side is
-    /// carried instead by the buy/sell luminance gap, which the achromatic
-    /// channel resolves several times smaller than the chromatic one, and by
-    /// [`side_offset`](Self::side_offset). The knob stays for anyone who
-    /// preferred the ring.
-    pub hollow_small_buys: bool,
-    /// Whether the fill is a flat disc or a shaded sphere.
-    pub render_mode: BubbleRenderMode,
-    /// How much a sphere-rendered bubble darkens toward its rim. Zero shades
-    /// nothing (the fill reads flat again); one pushes the rim to black.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub sphere_shading: f32,
-    /// Strength of the off-center light spot on a sphere-rendered bubble.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub sphere_highlight: f32,
-    /// How the full-size reference quantity is chosen.
-    pub size_reference: BubbleSizeReference,
-    /// Quantity mapped to [`max_radius`](Self::max_radius) when the reference
-    /// is [`BubbleSizeReference::Fixed`].
-    pub size_reference_quantity: f64,
-    /// Prints below this exact quantity are not drawn. Zero draws everything.
-    ///
-    /// Display-only and applied *after* liquidity association, so hiding small
-    /// prints never turns an aggression-aligned reduction into an
-    /// unattributed one.
-    pub min_quantity: f64,
-    /// Vertical separation, in pixels, between the two sides: buy bubbles are
-    /// nudged up (they lift the ask), sell bubbles down (they hit the bid).
-    ///
-    /// Deliberately unfaithful to the exact price — with a one-tick spread both
-    /// sides land on the same row and stack into an unreadable line. Zero
-    /// restores exact price placement.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub side_offset: f32,
-    /// How a bubble that ate resting liquidity says so.
-    ///
-    /// Replaces the older `show_consumption_front` switch. A preset written
-    /// before the crown existed simply does not carry this key, so it opens
-    /// wearing the crown — which is the point of changing the default, and is
-    /// covered by a test that such a preset still loads.
-    pub consumption_mark: ConsumptionMark,
-    /// Width of the vertical front, in pixels. Unused by
-    /// [`ConsumptionMark::Crown`], whose width is a proportion of the radius.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub front_width: f32,
-    /// Half-length of the front as a multiple of the bubble radius.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub front_length_scale: f32,
-    /// Whether a consuming bubble draws a ring around its rim.
-    pub show_impact_ring: bool,
-    /// Width of that ring, in pixels.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub impact_ring_width: f32,
-    /// Length of the consumption trail leaking to the right, in pixels. Zero
-    /// draws no trail.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub trail_length: f32,
-    /// Alpha of the trail at the bubble's edge; it fades to nothing.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub trail_opacity: f32,
-    /// Whether large bubbles print their quantity.
-    pub show_quantity_labels: bool,
-    /// Whether large bubbles print how many trades they cluster.
-    pub show_trade_count: bool,
-    /// Smallest radius that gets a label at all.
-    #[serde(serialize_with = "serialize_short_f32")]
-    pub label_min_radius: f32,
-    /// Buy-side colour override; absent follows the theme.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub buy_color: Option<[u8; 3]>,
-    /// Sell-side colour override; absent follows the theme.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sell_color: Option<[u8; 3]>,
-    /// Consumption-front colour override; absent follows the theme.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub front_color: Option<[u8; 3]>,
-    /// Trail colour override; absent follows the front colour.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub trail_color: Option<[u8; 3]>,
-    /// Label colour override; absent follows the theme.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub label_color: Option<[u8; 3]>,
-}
-
-impl Default for BubbleStyle {
-    fn default() -> Self {
-        Self {
-            // Area stays quantity-proportional, and the four radii are rungs
-            // of one φ ladder off `max_radius` — see `PHI`.
-            min_radius: DEFAULT_BUBBLE_MIN_RADIUS,
-            max_radius: DEFAULT_BUBBLE_MAX_RADIUS,
-            opacity: DEFAULT_BUBBLE_OPACITY,
-            // Zero, not a hairline: a sub-pixel stroke is antialiased into a
-            // grey fringe around every print, and in sphere mode the darkened
-            // rim already is the boundary. A second edge on top of it is what
-            // made each bubble wear three outlines.
-            outline_width: 0.0,
-            halo_strength: 0.1,
-            detail_min_radius: DEFAULT_DETAIL_MIN_RADIUS,
-            readable_min_radius: DEFAULT_READABLE_MIN_RADIUS,
-            hollow_small_buys: false,
-            // Shading is what keeps piled bubbles countable where their rims
-            // coincide, which a dense tape does constantly.
-            render_mode: BubbleRenderMode::Sphere,
-            sphere_shading: DEFAULT_SPHERE_SHADING,
-            sphere_highlight: DEFAULT_SPHERE_HIGHLIGHT,
-            size_reference: BubbleSizeReference::VisibleP99,
-            size_reference_quantity: 100.0,
-            min_quantity: 0.0,
-            // Wide enough that two prints on one price row visibly straddle
-            // it, which is the second channel carrying side at speck size.
-            side_offset: 3.0,
-            consumption_mark: ConsumptionMark::Crown,
-            front_width: 1.6,
-            front_length_scale: DEFAULT_FRONT_LENGTH_SCALE,
-            // The crown says "ate the book" without closing a second circle
-            // around the rim; a closed ring on top of it restates the same
-            // fact in the one shape that blurs the disc's own edge.
-            show_impact_ring: false,
-            impact_ring_width: 1.4,
-            // Off. The trail fired on nearly every print, so it carried almost
-            // no information, and it painted over the strip of heat map that
-            // answers the question it raised — did the level refill? The knob
-            // stays; the default stops smearing the book.
-            trail_length: 0.0,
-            trail_opacity: 0.45,
-            show_quantity_labels: true,
-            show_trade_count: true,
-            // Only the largest bubbles get a (text-layout-costly) label: the
-            // top of the radius range, not a number picked beside it.
-            label_min_radius: DEFAULT_LABEL_MIN_RADIUS,
-            buy_color: None,
-            sell_color: None,
-            front_color: None,
-            trail_color: None,
-            label_color: None,
-        }
-    }
-}
-
-impl BubbleStyle {
-    /// Clamp every numeric field to a value that is safe for geometry and math.
-    pub fn sanitize(&mut self) {
-        self.min_radius = finite_clamp(
-            self.min_radius,
-            0.5,
-            MAX_BUBBLE_MIN_RADIUS,
-            DEFAULT_BUBBLE_MIN_RADIUS,
-        );
-        self.max_radius = finite_clamp(
-            self.max_radius,
-            self.min_radius,
-            MAX_BUBBLE_MAX_RADIUS,
-            DEFAULT_BUBBLE_MAX_RADIUS,
-        );
-        self.opacity = finite_clamp(self.opacity, 0.05, 1.0, DEFAULT_BUBBLE_OPACITY);
-        self.outline_width = finite_clamp(self.outline_width, 0.0, 6.0, 0.0);
-        self.halo_strength = finite_clamp(self.halo_strength, 0.0, 1.0, 0.1);
-        self.detail_min_radius =
-            finite_clamp(self.detail_min_radius, 0.0, 32.0, DEFAULT_DETAIL_MIN_RADIUS);
-        self.readable_min_radius = finite_clamp(
-            self.readable_min_radius,
-            0.0,
-            MAX_READABLE_MIN_RADIUS,
-            DEFAULT_READABLE_MIN_RADIUS,
-        );
-        self.sphere_shading = finite_clamp(self.sphere_shading, 0.0, 1.0, DEFAULT_SPHERE_SHADING);
-        self.sphere_highlight =
-            finite_clamp(self.sphere_highlight, 0.0, 1.0, DEFAULT_SPHERE_HIGHLIGHT);
-        if !self.size_reference_quantity.is_finite() || self.size_reference_quantity <= 0.0 {
-            self.size_reference_quantity = 100.0;
-        }
-        if !self.min_quantity.is_finite() || self.min_quantity < 0.0 {
-            self.min_quantity = 0.0;
-        }
-        self.side_offset = finite_clamp(self.side_offset, 0.0, 40.0, 3.0);
-        self.front_width = finite_clamp(self.front_width, 0.5, 12.0, 1.6);
-        self.front_length_scale = finite_clamp(
-            self.front_length_scale,
-            0.2,
-            8.0,
-            DEFAULT_FRONT_LENGTH_SCALE,
-        );
-        self.impact_ring_width = finite_clamp(self.impact_ring_width, 0.5, 8.0, 1.4);
-        self.trail_length = finite_clamp(self.trail_length, 0.0, 120.0, 0.0);
-        self.trail_opacity = finite_clamp(self.trail_opacity, 0.0, 1.0, 0.45);
-        self.label_min_radius =
-            finite_clamp(self.label_min_radius, 4.0, 64.0, DEFAULT_LABEL_MIN_RADIUS);
-    }
-
-    /// Quantity at which a print stops being dust: the exact quantity whose
-    /// bubble lands on [`readable_min_radius`](Self::readable_min_radius).
-    ///
-    /// Inverts the renderer's area mapping — `radius = sqrt(min² + size² ·
-    /// (max² − min²))` at `size² = quantity / reference` — so the threshold
-    /// follows whatever radius range is configured instead of pinning a second
-    /// magic number beside it. `None` means nothing can be dust: no reference
-    /// to size against, or a floor already at the smallest drawn radius.
-    #[must_use]
-    pub fn dust_quantity(&self, reference: Decimal) -> Option<Decimal> {
-        if reference <= Decimal::ZERO || self.readable_min_radius <= self.min_radius {
-            return None;
-        }
-        let span = self.max_radius.powi(2) - self.min_radius.powi(2);
-        if span <= 0.0 {
-            return None;
-        }
-        let size_sq = (self.readable_min_radius.powi(2) - self.min_radius.powi(2)) / span;
-        Decimal::from_f32(size_sq.clamp(0.0, 1.0)).map(|share| reference * share)
-    }
-
-    /// Exact quantity below which a print is not drawn, if any.
-    #[must_use]
-    pub fn min_quantity_decimal(&self) -> Option<Decimal> {
-        (self.min_quantity > 0.0)
-            .then(|| Decimal::from_f64_retain(self.min_quantity))
-            .flatten()
-    }
-
-    /// Explicit full-size quantity, when the reference is fixed.
-    #[must_use]
-    pub fn fixed_reference_decimal(&self) -> Option<Decimal> {
-        Decimal::from_f64_retain(self.size_reference_quantity)
-            .filter(|value| *value > Decimal::ZERO)
-    }
-}
-
 fn finite_clamp(value: f32, low: f32, high: f32, fallback: f32) -> f32 {
     if value.is_finite() {
         value.clamp(low, high)
     } else {
         fallback.clamp(low, high)
-    }
-}
-
-/// How much market time the tape shows.
-///
-/// One number, one owner. The lane is a fixed band of screen, so its window is
-/// what decides how fast a print crosses it — and there is exactly one way to
-/// ask for that, phrased in one of two languages:
-///
-/// - [`Auto`](Self::Auto) follows the bars: the tape holds roughly one bar's
-///   worth of flow whatever the instrument or the bar type, and the zoom
-///   scales that reference. This is what the lane has always done, and it
-///   stays the default.
-/// - [`Fixed`](Self::Fixed) pins an amount of market time and ignores what the
-///   bars are doing. On an activity-sampled series a burst closes bars faster
-///   than a human reads prints, and a trader who wants the last two minutes in
-///   front of them is asking for two minutes, not for a multiplier.
-///
-/// The gesture over the tape edits whichever language is in force — the zoom
-/// in `Auto`, the milliseconds in `Fixed` — so dragging never silently
-/// switches modes.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum LaneWindow {
-    /// Follows the recent bars' typical duration, scaled by `zoom`.
-    Auto {
-        /// `1.0` fits about one bar's worth of market time in the band. Above
-        /// one shows less time (prints run across faster and further apart),
-        /// below one shows more (they crowd together and cluster).
-        zoom: f32,
-    },
-    /// A fixed amount of market time, whatever the bars do.
-    Fixed {
-        /// Exchange milliseconds shown in the band.
-        ms: i64,
-    },
-}
-
-impl Default for LaneWindow {
-    fn default() -> Self {
-        Self::Auto {
-            zoom: DEFAULT_LIVE_LANE_ZOOM,
-        }
-    }
-}
-
-impl LaneWindow {
-    /// Clamp into a range the renderer can use.
-    pub fn sanitize(&mut self) {
-        match self {
-            Self::Auto { zoom } => {
-                *zoom = finite_clamp(
-                    *zoom,
-                    MIN_LIVE_LANE_ZOOM,
-                    MAX_LIVE_LANE_ZOOM,
-                    DEFAULT_LIVE_LANE_ZOOM,
-                );
-            }
-            Self::Fixed { ms } => {
-                *ms = (*ms).clamp(MIN_LIVE_LANE_WINDOW_MS, MAX_LIVE_LANE_WINDOW_MS);
-            }
-        }
-    }
-
-    /// Market time the lane shows, given the automatic reference — the recent
-    /// bars' typical duration ([`reserved_span_ms`](super::reserved_span_ms)).
-    ///
-    /// Both modes end in the same clamp: a session-long reference cannot turn
-    /// the tape into a second chart, and a burst cannot make it an instant
-    /// wide, whichever language asked for the window.
-    #[must_use]
-    pub fn resolve_ms(self, reference_ms: i64) -> i64 {
-        let window = match self {
-            Self::Auto { zoom } => {
-                let zoom = if zoom.is_finite() {
-                    zoom.clamp(MIN_LIVE_LANE_ZOOM, MAX_LIVE_LANE_ZOOM)
-                } else {
-                    DEFAULT_LIVE_LANE_ZOOM
-                };
-                (reference_ms.max(1) as f64 / f64::from(zoom)).round() as i64
-            }
-            Self::Fixed { ms } => ms,
-        };
-        window.clamp(MIN_LIVE_LANE_WINDOW_MS, MAX_LIVE_LANE_WINDOW_MS)
-    }
-
-    /// How the resolved window compares to the automatic reference.
-    ///
-    /// The factor the clustering window is divided by, so a cluster covers a
-    /// constant *distance on screen* rather than a constant amount of time. In
-    /// `Auto` it works out to the zoom itself; in `Fixed` it is whatever the
-    /// pinned window is worth against the same reference — which is what keeps
-    /// a two-minute tape from drawing as a smear.
-    #[must_use]
-    fn cluster_scale(self, reference_ms: i64) -> f64 {
-        reference_ms.max(1) as f64 / self.resolve_ms(reference_ms).max(1) as f64
-    }
-
-    /// Apply a multiplicative zoom gesture, in whichever language is in force.
-    ///
-    /// `> 1` shows less market time in the same band, `< 1` more. A gesture
-    /// never changes the mode: pinning two minutes and then scrolling gives a
-    /// different fixed window, not a return to following the bars.
-    pub fn zoom_by(&mut self, factor: f32) {
-        if !factor.is_finite() || factor <= 0.0 {
-            return;
-        }
-        match self {
-            Self::Auto { zoom } => *zoom *= factor,
-            Self::Fixed { ms } => {
-                *ms = ((*ms as f64 / f64::from(factor)).round() as i64)
-                    .clamp(MIN_LIVE_LANE_WINDOW_MS, MAX_LIVE_LANE_WINDOW_MS);
-            }
-        }
-    }
-}
-
-/// A duration as a trader reads it: `8 s`, `1 min`, `1 min 30 s`, `250 ms`.
-///
-/// Sub-second windows keep their milliseconds because that is the only place
-/// the number carries information; past a minute the seconds are dropped when
-/// they are zero, so a preset reads as the round number it was chosen as.
-#[must_use]
-pub fn format_window_ms(ms: i64) -> String {
-    let ms = ms.max(0);
-    if ms < 1_000 {
-        return format!("{ms} ms");
-    }
-    let seconds = ms / 1_000;
-    let (minutes, seconds) = (seconds / 60, seconds % 60);
-    match (minutes, seconds) {
-        (0, seconds) => format!("{seconds} s"),
-        (minutes, 0) => format!("{minutes} min"),
-        (minutes, seconds) => format!("{minutes} min {seconds} s"),
-    }
-}
-
-/// Share of the lane's own window a print must fall behind before the axis
-/// says so.
-///
-/// Below it the newest bubble is a couple of pixels off the edge — saying
-/// anything would be noise on a tape that is, to the eye, current. The
-/// threshold is a share rather than a duration because the lane's window is
-/// the trader's own setting: what reads as "behind" on a four-second tape is
-/// invisible on a two-minute one.
-///
-/// A sixth of the tape, because the caption has to earn its appearance. A
-/// tighter threshold would light it up every time an ordinary market went
-/// quiet for a moment, and a warning that cries wolf during lunch is one the
-/// trader stops reading before the session that matters.
-const LANE_LAG_LABEL_SHARE: f64 = 1.0 / 6.0;
-
-/// Shortest silence the axis will call a silence, whatever the window.
-///
-/// The share alone collapses at the small end: the tape zooms to
-/// [`MIN_LIVE_LANE_WINDOW_MS`] of 200 ms, where a sixth is 33 ms — under the
-/// interval between two ordinary prints (a WIN recording measures 20 ms at the
-/// median and 285 ms at the worst). The caption would then be lit permanently
-/// on a tape that is perfectly current, which is the cry-wolf failure the
-/// share was chosen to avoid, reached from the other direction. Below a second
-/// there is no delay a person reads as one.
-const MIN_LANE_LAG_MS: i64 = 1_000;
-
-/// What the axis under the tape says about how old its newest mark is, or
-/// `None` while the tape is current enough that saying anything is noise.
-///
-/// The caller decides *whether to ask*: this answers about the aggression
-/// marks, so a pane drawing only the depth map has no missing bubbles to
-/// explain and must not be handed an age at all.
-///
-/// The lane's right edge is the newer of the book clock and the print clock,
-/// and only prints draw bubbles. So a book running ahead of the tape — a quiet
-/// stretch, or prints held up on their way here — puts every bubble left of
-/// the edge, and past the whole window puts none of them on the tape at all.
-/// An empty tape reads as "nothing is trading", which is the one thing it
-/// never means: the prints are on the chart, in their bar's slot. The axis is
-/// where that gets said, because it is already the readout for what the lane
-/// is showing.
-#[must_use]
-pub fn lane_lag_label(window_ms: i64, tape_age: Option<TapeAge>) -> Option<String> {
-    // Nothing has printed since the tape started watching. There is no mark to
-    // be late, so the wording never invites the reader to look for one, and no
-    // window threshold applies — an empty tape with a running book is worth
-    // saying at any zoom.
-    let age = match tape_age? {
-        TapeAge::NothingYet(watched) => {
-            return Some(format!("no print for {}", format_window_ms(watched)));
-        }
-        TapeAge::Behind(age) => age,
-    };
-    let window = window_ms.max(1);
-    #[allow(clippy::cast_precision_loss)]
-    let share_floor = (window as f64 * LANE_LAG_LABEL_SHARE) as i64;
-    if age <= share_floor.max(MIN_LANE_LAG_MS) {
-        return None;
-    }
-    // Past the window there is no bubble left on the tape to be late: the
-    // wording changes with it, because "last print 40 s ago" beside an empty
-    // tape still invites the reader to look for the mark it describes.
-    //
-    // Strictly past. The lane spans `[now - window, now]`, so at exactly the
-    // window the newest mark is still drawn, on the tape's leftmost pixel —
-    // telling the reader there is nothing to look for while it is on screen
-    // is the same lie in the other direction.
-    if age > window {
-        return Some(format!("no print for {}", format_window_ms(age)));
-    }
-    Some(format!("last print {} back", format_window_ms(age)))
-}
-
-/// How a tape window reads in a menu.
-///
-/// `reference_ms` is the automatic reference, when the caller knows it. The
-/// automatic entry states the duration it currently works out to whenever it
-/// is supplied: "auto" alone names a policy, and a trader choosing between it
-/// and `30 s` is comparing durations, so the one number that would let them
-/// compare has no business being the one number hidden.
-#[must_use]
-pub fn lane_window_label(window: LaneWindow, reference_ms: Option<i64>) -> String {
-    match window {
-        LaneWindow::Auto { .. } => reference_ms.map_or_else(
-            || "auto".to_owned(),
-            |reference| {
-                format!(
-                    "auto (≈ {})",
-                    format_window_ms(window.resolve_ms(reference))
-                )
-            },
-        ),
-        LaneWindow::Fixed { ms } => format_window_ms(ms),
-    }
-}
-
-/// Whether a menu entry names the window already in force.
-///
-/// Two `Auto`s match whatever their zoom, so selecting the automatic entry
-/// while already following the bars cannot quietly reset a zoom the trader
-/// set; two `Fixed`s have to name the same duration.
-#[must_use]
-pub fn same_lane_window(current: LaneWindow, option: LaneWindow) -> bool {
-    match (current, option) {
-        (LaneWindow::Auto { .. }, LaneWindow::Auto { .. }) => true,
-        (LaneWindow::Fixed { ms: current }, LaneWindow::Fixed { ms: option }) => current == option,
-        _ => false,
-    }
-}
-
-/// How the rolling tape past the last bar is drawn.
-///
-/// The live lane is a pane of its own, pinned to the right edge of the chart:
-/// history is compressed into equal-width bar slots that pan and zoom, the lane
-/// is a fixed band of screen showing a fixed window of market time that always
-/// ends at now. Being its own pane is what earns it settings of its own — how
-/// wide it is, how much time fits in it, how big its bubbles are, and which of
-/// the two flow layers it draws, none of which the candles beside it get a say
-/// in.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(from = "LiveLaneStyleRepr", into = "LiveLaneStyleRepr")]
-pub struct LiveLaneStyle {
-    /// Width of the lane as a share of the chart.
-    ///
-    /// Measured against the chart rather than the candle so that the tape is a
-    /// *place* on screen: zooming the time axis changes how many bars fit
-    /// beside it, never how much room it gets. Fixed, and so is the market time
-    /// it shows, which together give the tape a constant pixels-per-ms rate —
-    /// a print enters at the right edge and slides left at the same speed. A
-    /// bar closing changes neither, so nothing on the tape ever jumps.
-    pub width_share: f32,
-    /// How much market time fits in the band: following the bars, or pinned.
-    ///
-    /// The clustering window follows whatever this resolves to
-    /// ([`effective_cluster_ms`](Self::effective_cluster_ms)), which is what
-    /// turns a crowded tape into fewer, bigger marks instead of a smear.
-    pub window: LaneWindow,
-    /// Clustering window applied to prints inside the lane, before the window
-    /// scales it.
-    ///
-    /// `None` inherits [`HeatmapConfig::bubble_cluster_ms`]. A shorter window
-    /// than history's buys detail where there is room for it; a longer one
-    /// gathers a fast tape into readable marks.
-    pub cluster_ms: Option<i64>,
-    /// Multiplier applied to both bubble radii inside the lane.
-    pub radius_scale: f32,
-    /// Whether the lane's boundary and the live-edge line are drawn.
-    pub show_marks: bool,
-    /// Whether the tape is on the canvas at all.
-    ///
-    /// Off, the lane reserves no width ([`Self::resolved_width_px`] answers
-    /// zero), so the candles take the whole canvas and nothing downstream —
-    /// divider, projection, marks — has a band to work in. The two layer
-    /// switches below keep whatever they were set to while it is off: turning
-    /// the tape back on must return the tape that was switched off, not a
-    /// default one.
-    pub enabled: bool,
-    /// Whether the depth map is drawn *on the tape*.
-    ///
-    /// Each pane answers for its own canvas, and the tape answers with a value
-    /// of its own rather than by following the candles: the compressed history
-    /// is read one way and the rolling tape another, so a trader who clears the
-    /// candles has not thereby asked for the book to vanish where the book is
-    /// actually being read. On by default — the tape's whole job is to show the
-    /// book and the prints landing into it.
-    pub show_depth: bool,
-    /// Whether aggression bubbles are drawn *on the tape*. Same rule as
-    /// [`show_depth`](Self::show_depth), and on by default for the same reason.
-    pub show_aggressions: bool,
-}
-
-impl Default for LiveLaneStyle {
-    fn default() -> Self {
-        Self {
-            width_share: DEFAULT_LIVE_LANE_SHARE,
-            window: LaneWindow::default(),
-            cluster_ms: None,
-            radius_scale: DEFAULT_LIVE_LANE_RADIUS_SCALE,
-            show_marks: true,
-            enabled: true,
-            show_depth: true,
-            show_aggressions: true,
-        }
-    }
-}
-
-/// What a [`LiveLaneStyle`] looks like on disk.
-///
-/// The window is one field in memory and two on disk, and the split is
-/// deliberate: `time_zoom` is the name every preset written before this mode
-/// existed already uses, so reading one back has to find it where it was left.
-/// `window_ms` is absent from all of them, which is exactly the sentence "this
-/// lane follows the bars" — the mode a file from before the choice existed is
-/// in.
-///
-/// Nothing outside serialization sees this type. [`LaneWindow`] stays the only
-/// owner of the number in memory, so no two fields can disagree about how much
-/// market time the tape is showing.
-///
-/// The two layer switches are `Option<bool>` here and plain `bool` in memory,
-/// and that is the whole of the compatibility story. A file written while the
-/// tape still inherited the candles' switches left the key out whenever nobody
-/// had chosen, which read as "follow the chart"; there is no chart to follow
-/// any more, so an absent answer becomes the tape's own default — on. A file
-/// that says `false` said it on purpose and is still obeyed.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-struct LiveLaneStyleRepr {
-    width_share: f32,
-    time_zoom: f32,
-    window_ms: Option<i64>,
-    cluster_ms: Option<i64>,
-    radius_scale: f32,
-    show_marks: bool,
-    enabled: bool,
-    show_depth: Option<bool>,
-    show_aggressions: Option<bool>,
-}
-
-impl Default for LiveLaneStyleRepr {
-    fn default() -> Self {
-        Self::from(LiveLaneStyle::default())
-    }
-}
-
-impl From<LiveLaneStyleRepr> for LiveLaneStyle {
-    fn from(repr: LiveLaneStyleRepr) -> Self {
-        Self {
-            width_share: repr.width_share,
-            // A pinned window is the explicit choice; without one the file is
-            // asking to follow the bars, at whatever zoom it recorded.
-            window: repr.window_ms.map_or(
-                LaneWindow::Auto {
-                    zoom: repr.time_zoom,
-                },
-                |ms| LaneWindow::Fixed { ms },
-            ),
-            cluster_ms: repr.cluster_ms,
-            radius_scale: repr.radius_scale,
-            show_marks: repr.show_marks,
-            enabled: repr.enabled,
-            // Absent, or the `null` an inheriting file wrote: the tape's own
-            // default. See `LiveLaneStyleRepr`.
-            show_depth: repr.show_depth.unwrap_or(true),
-            show_aggressions: repr.show_aggressions.unwrap_or(true),
-        }
-    }
-}
-
-impl From<LiveLaneStyle> for LiveLaneStyleRepr {
-    fn from(style: LiveLaneStyle) -> Self {
-        // Pinning a window parks the zoom at its default rather than keeping
-        // the last one: "automatic" means the bars decide, and a remembered
-        // multiplier waiting to reappear is not that.
-        let (time_zoom, window_ms) = match style.window {
-            LaneWindow::Auto { zoom } => (zoom, None),
-            LaneWindow::Fixed { ms } => (DEFAULT_LIVE_LANE_ZOOM, Some(ms)),
-        };
-        Self {
-            width_share: style.width_share,
-            time_zoom,
-            window_ms,
-            cluster_ms: style.cluster_ms,
-            radius_scale: style.radius_scale,
-            show_marks: style.show_marks,
-            enabled: style.enabled,
-            show_depth: Some(style.show_depth),
-            show_aggressions: Some(style.show_aggressions),
-        }
-    }
-}
-
-impl LiveLaneStyle {
-    /// Clamp every value into a range the renderer can use.
-    pub fn sanitize(&mut self) {
-        self.width_share = finite_clamp(
-            self.width_share,
-            MIN_LIVE_LANE_SHARE,
-            MAX_LIVE_LANE_SHARE,
-            DEFAULT_LIVE_LANE_SHARE,
-        );
-        self.window.sanitize();
-        self.radius_scale = finite_clamp(
-            self.radius_scale,
-            MIN_LIVE_LANE_RADIUS_SCALE,
-            MAX_LIVE_LANE_RADIUS_SCALE,
-            DEFAULT_LIVE_LANE_RADIUS_SCALE,
-        );
-        self.cluster_ms = self
-            .cluster_ms
-            .map(|window| window.clamp(0, MAX_BUBBLE_CLUSTER_MS));
-    }
-
-    /// Lane width in pixels for a chart this wide.
-    ///
-    /// The lane is a pane, not a number of candle slots: it takes
-    /// `chart_width × width_share` of screen whatever the candles are doing, so
-    /// zooming the time axis changes how many bars fit beside the tape and
-    /// never how much room the tape gets.
-    #[must_use]
-    pub fn resolved_width_px(&self, chart_width: f32) -> f32 {
-        // The one gate the whole switch hangs off. A tape that is off reserves
-        // no band, and everything downstream already reads a zero width as
-        // "there is no lane": no divider, no lane rungs, no drawing boundary,
-        // no projection asked for on the tape's account.
-        if !self.enabled {
-            return 0.0;
-        }
-        if !chart_width.is_finite() || chart_width <= 0.0 {
-            return 0.0;
-        }
-        let share = if self.width_share.is_finite() {
-            self.width_share
-                .clamp(MIN_LIVE_LANE_SHARE, MAX_LIVE_LANE_SHARE)
-        } else {
-            DEFAULT_LIVE_LANE_SHARE
-        };
-        // The floor yields to the cap on a narrow window: half a small chart
-        // beats a fixed band that would eat all of it.
-        let floor = MIN_LIVE_LANE_WIDTH_PX.min(chart_width * MAX_LIVE_LANE_SHARE);
-        (chart_width * share).max(floor)
-    }
-
-    /// Market time the lane shows, given the automatic reference window (the
-    /// recent bars' typical duration).
-    #[must_use]
-    pub fn window_ms(&self, reference_ms: i64) -> i64 {
-        self.window.resolve_ms(reference_ms)
-    }
-
-    /// Clustering window for prints inside the lane, given history's own and
-    /// the automatic reference the lane's window is measured against.
-    ///
-    /// Scaled by the same factor as the window itself, so a cluster covers a
-    /// constant *distance on screen* rather than a constant amount of time:
-    /// a wider window gathers the crowd it creates into fewer, bigger bubbles
-    /// instead of piling prints on top of each other. The factor comes from
-    /// the resolved window rather than from the zoom, which is what lets a
-    /// pinned two-minute tape cluster like the zoomed-out tape it is.
-    #[must_use]
-    pub fn effective_cluster_ms(&self, history_cluster_ms: i64, reference_ms: i64) -> i64 {
-        let base = self.cluster_ms.unwrap_or(history_cluster_ms).max(0);
-        if base == 0 {
-            // One bubble per print, at every window: raw is a choice, not a
-            // size.
-            return 0;
-        }
-        let scaled = (base as f64 / self.window.cluster_scale(reference_ms)).round() as i64;
-        scaled.clamp(1, MAX_BUBBLE_CLUSTER_MS)
-    }
-
-    /// Bubble radius range inside the lane, given the shared bubble style.
-    #[must_use]
-    pub fn scaled_radii(&self, bubbles: &BubbleStyle) -> (f32, f32) {
-        let scale = if self.radius_scale.is_finite() {
-            self.radius_scale
-                .clamp(MIN_LIVE_LANE_RADIUS_SCALE, MAX_LIVE_LANE_RADIUS_SCALE)
-        } else {
-            DEFAULT_LIVE_LANE_RADIUS_SCALE
-        };
-        let min = (bubbles.min_radius * scale).clamp(0.0, MAX_BUBBLE_MIN_RADIUS);
-        let max = (bubbles.max_radius * scale).clamp(MIN_BUBBLE_MAX_RADIUS, MAX_BUBBLE_MAX_RADIUS);
-        (min.min(max), max)
     }
 }
 
@@ -1523,7 +514,9 @@ impl HeatmapConfig {
 
 #[cfg(test)]
 mod tests {
+    use super::bubbles::SERIALIZED_FLOAT_PLACES;
     use super::*;
+    use crate::history::TapeAge;
 
     #[test]
     fn defaults_are_off_and_bounded() {
